Split endpoint trace graph helpers
